correção de erros sem login

+++ pages/views.py
@@ -86,12 +86,14 @@ def getName(id):
 def home_view(request, *args, **kwargs):
     current_user = request.user
     nmec = current_user.username
-    sub = AsSubdject.objects.filter(user=nmec)
-    sub = [s.subject for s in sub]
     subjects = {}
-    for s in sub:
-        subjects[s] = getName(s)
-    print(subjects)
+    if nmec:
+        sub = AsSubdject.objects.filter(user=nmec)
+        sub = [s.subject for s in sub]
+
+        for s in sub:
+            subjects[s] = getName(s)
+        print(subjects)
     params = {
         'subjects': subjects
     }
@@ -107,21 +109,23 @@ def subjects(request, *args, ** kwargs):
 
     current_user = request.user
     nmec = current_user.username
-    sub = AsSubdject.objects.filter(user=nmec)
-    sub = [s.subject for s in sub]
-    # AsSubdject.objects.filter(user=88753, subject='ihc').delete()
-    if request.method == 'POST':
-        selected_subjects = request.POST.getlist('subjects')
-        for s in selected_subjects:
-            print(s)
-            if s not in sub:
-                new_sub = AsSubdject(user=nmec, subject=s)
-                new_sub.save()
-                sub.append(s)
-        for s in sub:
-            if (s in first or s in second) and s not in selected_subjects:
-                AsSubdject.objects.get(user=nmec, subject=s).delete()
-                sub.remove(s)
+    sub = []
+    if nmec:
+        sub = AsSubdject.objects.filter(user=nmec)
+        sub = [s.subject for s in sub]
+        # AsSubdject.objects.filter(user=88753, subject='ihc').delete()
+        if request.method == 'POST':
+            selected_subjects = request.POST.getlist('subjects')
+            for s in selected_subjects:
+                print(s)
+                if s not in sub:
+                    new_sub = AsSubdject(user=nmec, subject=s)
+                    new_sub.save()
+                    sub.append(s)
+            for s in sub:
+                if (s in first or s in second) and s not in selected_subjects:
+                    AsSubdject.objects.get(user=nmec, subject=s).delete()
+                    sub.remove(s)
 
     params = {'year': year,
               "first_sem": first,
